+++ package.json
@@ -4,7 +4,7 @@
   "description": "Shared CKEditor 4 GitHub workflows.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node tests/index.js"
   },
   "repository": {
     "type": "git",
