+++ tests/config.js
@@ -1,0 +1,26 @@
+const dotenv = require( 'dotenv' );
+
+function config() {
+	dotenv.config();
+
+	verifyEnvVariables( ['AUTH_KEY', 'OWNER', 'REPO' ] );
+}
+
+function verifyEnvVariables( requiredVariables ) {
+	let anyMissingVariable =  false;
+
+	requiredVariables.forEach( variable => {
+		if ( !process.env[variable] ) {
+			console.log( chalk.red( `Missing ${variable} env variable!` ) );
+			if ( !anyMissingVariable ) {
+				anyMissingVariable = true;
+			}
+		}
+	} );
+
+	if( anyMissingVariable ) {
+		process.exit( -1 );
+	}
+}
+
+module.exports = config;
